requirements.txt
python-telegram-bot==20.3
Flask==2.3.3
langdetect==1.0.9
Pillow
python-dotenv
